+++ src/superposition/game/system/PlayerController.scala
@@ -6,7 +6,7 @@ import com.badlogic.gdx.Gdx.input
 import com.badlogic.gdx.Input.Keys
 import com.badlogic.gdx.Input.Keys.SPACE
 import superposition.game.component._
-import superposition.game.system.PlayerControl.{carryGate, updateCarriedPositions, updatePlayerPosition, walk}
+import superposition.game.system.PlayerController.{carryGate, updateCarriedPositions, updatePlayerPosition, walk}
 import superposition.math.{Vector2d, Vector2i}
 import superposition.quantum.{Gate, Translate, X}
 
@@ -14,7 +14,7 @@ import scala.Function.const
 import scala.jdk.CollectionConverters._
 import scala.math.exp
 
-final class PlayerControl
+final class PlayerController
   extends IteratingSystem(Family.all(classOf[Player], classOf[Position], classOf[Quantum]).get) {
   private var carryables: Iterable[Entity] = Nil
 
@@ -36,7 +36,7 @@ final class PlayerControl
   }
 }
 
-private object PlayerControl {
+private object PlayerController {
   private val WalkKeys: Map[Int, Vector2d] = Map(
     Keys.W -> Vector2d(0, 1),
     Keys.A -> Vector2d(-1, 0),
